fix: harden credential redaction, cost-ceiling validation, and connection form submit
Preserve quote boundaries when redacting credentials so a redacted JSON document stays parseable. Validate cost_ceiling at AgentContext construction via a real from_identity path instead of an unvalidated model_copy. Bound cold-import fixture concurrency to cpu_count with a timeout so CI is not oversubscribed into MemoryError. Route secret capture errors through getErrorCode, guard in-flight connection-type fetches against reset races, and split the connection submit chain into connection-submit.ts to stay within module-size budgets. Correct the codespell contraction and align audit-chain and controller tests with the credential-handle create contract.
