+++ web/src/stores/connections.ts
@@ -1,6 +1,9 @@
 import { create } from 'zustand'
 import { createCrudActions } from './connections/crud-actions'
-import { createListActions } from './connections/list-actions'
+import {
+  createListActions,
+  invalidateInFlightConnectionRequests,
+} from './connections/list-actions'
 import type { ConnectionsState } from './connections/types'
 
 export type { ConnectionsState } from './connections/types'
@@ -26,5 +29,10 @@ export const useConnectionsStore = create<ConnectionsState>()((set, get) => ({
   ...INITIAL_STATE,
   ...createListActions(set, get),
   ...createCrudActions(set, get),
-  reset: () => set({ ...INITIAL_STATE }),
+  reset: () => {
+    // Invalidate any in-flight list/type fetch first so a late response cannot
+    // repopulate the store after it is cleared.
+    invalidateInFlightConnectionRequests()
+    set({ ...INITIAL_STATE })
+  },
 }))
